Added SimpleWrapper script and removed the other wrapper.
