+++ scripts/Makefile
@@ -9,8 +9,6 @@ all:
 install:
 	install -d $(PREFIX)/bin
 	install -m 755 spectral_indices.py $(PREFIX)/bin
-	install -m 755 Cmd.py $(PREFIX)/bin
-	install -m 755 ScriptHelper.py $(PREFIX)/bin
 
 
 clean:
