+++ scripts/spectral_indices.py
@@ -1,205 +1,122 @@
 #! /usr/bin/env python
-'''
------     -----     -----     -----     -----     -----     -----     -----
-Project Name: SI (Spectral Indices)
-Source: https://github.com/USGS-EROS/espa-spectral-indices
-License: NASA Open Source Agreement 1.3
-USGS Designation: EROS Science Processing Architecture (ESPA)
-Description:
-    Spectral Indices produces the desired spectral index products for the
-    input surface reflectance or TOA reflectance bands. The options
-    include: NDVI, EVI, SAVI, MSAVI, NDMI (also known as NDWI or NDII),
-    NBR, and NBR2. The user may specify one, some, or all of the
-    supported indices for output.
-
-Script:
-    File: spectral_indices.py
-    Author: ngenetzky@usgs.gov
-    Purpose: Master script for creating spectral indices products.
-    Help/Usage/Version: See "spectral_indices.py --help"
-    Note: Script is based on source code from: do_spectral_indices.py
-        From the project: github.com/USGS-EROS/espa-spectral-indices/
-        Written by: gschmidt@usgs.gov
-
-Executables:
-    File: spectral_indices
-    Author:  gschmidt@usgs.gov
-    Help/Usage/Version: See "spectral_indices.py --help"
------     -----     -----     -----     -----     -----     -----     -----
-'''
-import sys
+import argparse
+import os
 import logging
-from Cmd import Cmd
-from ScriptHelper import ScriptHelper, ScriptArgParser
+import sys
+import commands
 
 
 def get_logger():
+        '''Replicated from ESPA logger module'''
+        # Setup the Logger with the proper configuration
+        logging.basicConfig(format=('%(asctime)s.%(msecs)03d %(process)d'
+                                    ' %(levelname)-8s'
+                                    ' %(filename)s:%(lineno)d:'
+                                    '%(funcName)s -- %(message)s'),
+                            datefmt='%Y-%m-%d %H:%M:%S',
+                            level=logging.INFO)
+        return logging.getLogger(__name__)
 
-    # Setup the Logger with the proper configuration
-    logging.basicConfig(format=('%(asctime)s.%(msecs)03d %(process)d'
-                                ' %(levelname)-8s'
-                                ' %(filename)s:%(lineno)d:'
-                                '%(funcName)s -- %(message)s'),
-                        datefmt='%Y-%m-%d %H:%M:%S',
-                        level=logging.INFO)
-    return logging.getLogger(__name__)
+
+def parse_only_xml():
+        '''Will only parse --xml XML_FILENAME from cmdline.
+
+        Precondition:
+            '--xml FILENAME' exists in command line arguments
+        Postcondition:
+            returns xml_filename
+        '''
+        # Try to parse out the XML so the exe can be determined
+        parse_xml = argparse.ArgumentParser(add_help=False)
+        parse_xml.add_argument('--xml', action='store',
+                               dest='xml_filename', required=True,
+                               help='Input XML metadata file',
+                               metavar='FILE')
+        (temp, extra_args) = parse_xml.parse_known_args()
+        return temp.xml_filename
 
 
-class SI(ScriptHelper):  # ####################################################
-    '''Parse request, build command line and return output from executable.
-    '''
-    __title__ = 'Spectral Indices'
+def is_landsat8(xml_filename):
+        '''Reads xml_filename for satellitecode, checks if L8.
+        Precondition:
+            (1) satellite_code is the first 3 characters of input product id
+            (2) xml_filename is supplied as an argument
+        Postcondition:
+            returns True if this satellite_code is in ['lc8','lo8']
+        '''
+        satellite_code = xml_filename[0:3]
+        l8_prefixes = ['LC8', 'LO8']
+        if satellite_code in l8_prefixes:
+            return True
+        else:
+            return False
 
-    def __init__(self):
-        ScriptHelper.__init__(self)
-        self.title = 'Spectral Indices'
-        self.description = ('''
-        spectral_indices produces the desired spectral index products for the
-        input surface reflectance or TOA reflectance bands. The options
-        include: NDVI, EVI, SAVI, MSAVI, NDMI (also known as NDWI or NDII),
-        NBR, and NBR2. The user may specify one, some, or all of the
-        supported indices for output.
-        ''')
 
-    def build_parser(self):
-        '''See ScriptHelper.build_parser() for more details'''
-        self.parser = ScriptArgParser(description=self.description)
-        self.parser = ScriptHelper.add_common_args(self.parser, xml=True)
+class EXECUTE_ERROR(Exception):
+    '''Raised when command in execute_cmd returns with error'''
+    def __init__(self, message, *args):
+        self.message = message
+        Exception.__init__(self, message, *args)
 
-        # Additional parameters
-        self.parser.add_argument('--toa', dest='toa', default=False,
-                                 action='store_true',
-                                 help='process TOA bands'
-                                 ' instead of surface reflectance bands')
 
-        self.parser.add_argument('--ndvi', dest='ndvi', default=False,
-                                 action='store_true',
-                                 help='process NDVI'
-                                 '(normalized difference vegetation index)')
-
-        self.parser.add_argument('--ndmi', dest='ndmi', default=False,
-                                 action='store_true',
-                                 help='process NDMI'
-                                 '(normalized difference moisture index)')
-
-        self.parser.add_argument('--nbr', dest='nbr', default=False,
-                                 action='store_true',
-                                 help=('process NBR'
-                                       '(normalized burn ratio)'))
-
-        self.parser.add_argument('--nbr2', dest='nbr2', default=False,
-                                 action='store_true',
-                                 help='process NBR2'
-                                 '(normalized burn ratio2)')
-
-        self.parser.add_argument('--savi', dest='savi', default=False,
-                                 action='store_true',
-                                 help='process SAVI'
-                                 '(soil adjusted vegetation index)')
-
-        self.parser.add_argument('--msavi', dest='msavi', default=False,
-                                 action='store_true',
-                                 help='process modified SAVI'
-                                 '(soil adjusted vegetation index)')
-
-        self.parser.add_argument('--evi', dest='evi', default=False,
-                                 action='store_true',
-                                 help='process EVI'
-                                 '(enhanced vegetation index)')
-
-        # Common command line arguments
-        self.parser = ScriptHelper.add_common_args(self.parser,
-                                                   debug=False,
-                                                   verbose=True,
-                                                   version=True)
-
-    def check_product_requested(self):
-        '''Ensures that an SI product was requested
+def execute_cmd(cmd_string):
+        '''Execute a command line and return the terminal output
 
         Raises:
-            ScriptHelper.NO_ACTION_REQUESTED
+            EXECUTE_ERROR (Stdout/Stderr)
+
+        Returns:
+            output:The stdout and/or stderr from the executed command.
         '''
-        # TODO - Replace with argparse group
-        if (not self.args.ndvi and
-                not self.args.ndmi and
-                not self.args.nbr and
-                not self.args.nbr2 and
-                not self.args.savi and
-                not self.args.msavi and
-                not self.args.evi):
-            raise ScriptHelper.NO_ACTION_REQUESTED('No spectral indices'
-                                                   ' product specified to'
-                                                   ' be processed')
+        (status, output) = commands.getstatusoutput(cmd_string)
 
-    def build_cmd_line(self):
-        '''See ScriptHelper.build_cmd_line() for more details'''
-        self.check_common_prereq(check_xml=True)
-        self.check_product_requested()
+        if status < 0:
+            message = "Application terminated by signal [%s]" % cmd_string
+            if len(output) > 0:
+                message = ' Stdout/Stderr is: '.join([message, output])
+            raise EXECUTE_ERROR(message)
 
-        self.cmd = Cmd(self.get_exe_filename())
-        self.cmd.add_param('--xml', self.args.xml_filename)
+        if status != 0:
+            message = "Application failed to execute [%s]" % cmd_string
+            if len(output) > 0:
+                message = ' Stdout/Stderr is: '.join([message, output])
+            raise EXECUTE_ERROR(message)
 
-        if self.args.verbose:
-            self.cmd.add_param('--verbose')
-        if self.args.toa:
-            self.cmd.add_param('--toa')
-        if self.args.ndvi:
-            self.cmd.add_param('--ndvi')
-        if self.args.ndmi:
-            self.cmd.add_param('--ndmi')
-        if self.args.nbr:
-            self.cmd.add_param('--nbr')
-        if self.args.nbr2:
-            self.cmd.add_param('--nbr2')
-        if self.args.savi:
-            self.cmd.add_param('--savi')
-        if self.args.msavi:
-            self.cmd.add_param('--msavi')
-        if self.args.evi:
-            self.cmd.add_param('--evi')
+        if os.WEXITSTATUS(status) != 0:
+            message = ("Application [%s] returned error code [%d]"
+                       % (cmd_string, os.WEXITSTATUS(status)))
+            if len(output) > 0:
+                message = ' Stdout/Stderr is: '.join([message, output])
+            raise EXECUTE_ERROR(message)
 
-    def handle_exception(self):
-        '''Will handle any exception that the ScriptHelper understands.
+        return output
 
-        See ScriptHelper.handle_exception() for more details.
-        See sys.exc_info() = [type,value,traceback]
-        Description:
-        Return:
-            exceptionHandled: returns False if the exception was not understood
-        '''
-        e_type = sys.exc_info()[0]
-        if e_type is ScriptHelper.NO_ACTION_REQUESTED:
-            self.print_custom_help()
-            self.logger.warning('NO_ACTION_REQUESTED:No SI product specified.')
-            return True
-        return ScriptHelper.handle_exception(self)
 
-    def get_executables_version(self):
-        string_long_version = ScriptHelper.get_executables_version(self)
-        list_long_version = string_long_version.split(' ')
-        return list_long_version[2]  # Expect "NAME version x.x.x
-
-    def get_exe_filename(self):
-        return 'spectral_indices'
+def get_executable(isLandsat8):
+    return 'spectral_indices'
 
 
 def main():
     logger = get_logger()
 
-    script = SI()
+    xml_filename = parse_only_xml()
+    isLandsat8 = is_landsat8(xml_filename)
+
+    cmd = [get_executable(isLandsat8)]
+    cmd.extend(sys.argv[1:])  # Pass all arguments through
+    cmd_string = ' '.join(cmd)
     try:
-        script.run()
-    except Cmd.EXECUTE_ERROR as e:
+        logger.info('>>'+cmd_string)
+        output = execute_cmd(cmd_string)
+
+        if len(output) > 0:
+            logger.info("\n{0}".format(output))
+    except EXECUTE_ERROR:
         logger.exception(('Error running {0}.'
                           'Processing will terminate.'
-                          ).format(script.title))
-        try:
-            logger.info(e.args[0])
-        except NameError:  # No Message
-            pass
-        raise
+                          ).format(os.path.basename(__file__)))
+        raise  # Re-raise so exception message will be shown.
 
-script = None
 if __name__ == '__main__':
     main()
